requirements.txt
nltk==3.9.1
pandas==2.2.2
requests==2.32.3
spotipy==2.25.0
Flask==3.0.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
huggingface-hub==0.28.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
transformers==4.48.3
